requirements.txt
Flask>=3.0.0
flask-mongoengine>=1.0.0
mongoengine>=0.27.0
pymongo>=4.5.0
gunicorn>=21.0.0
python-dotenv>=1.0.0

